Isometric Translation, Cube fix

+++ include/Vector2.hpp
@@ -11,7 +11,7 @@ struct Vector2
 {
     float x;
     float y;
-    Vector2(int x = 0, int y = 0)
+    Vector2(float x = 0, float y = 0)
         : x(x), y(y)
     {
     }
